% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_divergence.R
\name{js_divergence}
\alias{js_divergence}
\title{Calculate Jensen-Shannon Divergence based on character n-grams}
\usage{
js_divergence(name_1, name_2, n = 2)
}
\arguments{
\item{name_1}{Character string}

\item{name_2}{Character string}

\item{n}{Integer length of character n-grams (default 2)}
}
\value{
Numeric scalar bounded [0, 1]
}
\description{
Calculate Jensen-Shannon Divergence based on character n-grams
}
\examples{
js_divergence("Jon Smith", "John Smith")
js_divergence("Jon Smith", "Elizabeth Howell")
}
